wygibasy w envelopie ampa

+++ libs/Audio/effect_envelope.h
@@ -84,7 +84,7 @@ public:
 	void clearEndReleaseFlag();
 	uint8_t getState();
 
-	void syncTrackerSeq(uint16_t val, float seqSpeed);
+	void syncTrackerSeq(uint32_t val, float seqSpeed);
 	void setSyncStartStep(uint16_t n);
 	void setPhaseNumbers(int8_t n1, int8_t n2);
 	void setSyncRate(float sync);
