+++ source/mtEnvelopeGenerator.h
@@ -59,7 +59,7 @@ public:
 
 	void putMidiParameter(uint8_t ch, uint8_t parNo, uint8_t value);
 
-	void syncTrackerSeq(uint16_t val, float seqSpeed);
+	void syncTrackerSeq(uint32_t val, float seqSpeed);
 	void setSyncStartStep(uint16_t n);
 	void setPhaseNumbers(int8_t n1, int8_t n2);
 	void setSyncRate(float sync);
